fix: modified website header to be sticky on scroll
Added a top margin the same height as the fixed header to the main content page. This prevents content to not start under the header.

Added a z-index property less than the header component so that the header is always above the content when scrolling.

Issue [Hacktoberfest]: Freeze navbar while scrolling

Fixes #24